+++ views/admin-tree/_form.php
@@ -12,6 +12,8 @@ use yii\widgets\ActiveForm;
 
 <?php $form = ActiveForm::begin(); ?>
 
+<?= $form->field($model, 'files')->widget(\skeeks\cms\widgets\formInputs\StorageImages::className())->label(false); ?>
+
 <?= $form->field($model, 'name')->textInput(['maxlength' => 255]) ?>
 
 <?= $form->field($model, 'type')->widget(